Fixes #463
When opening from Windows search we have issues when using Path.Combine where the first part of the path is a string (not an absolute path)

```
Path.Combine("StoredData", ...
```

would try to write to system32.

Going forwards we have to be careful when using Path.Combine, unless we can change current working directory?